Move VisibleDeprecationWarning
