--- hyperspy/misc/hspy_warnings.py
@@ -1,8 +1,0 @@
-class VisibleDeprecationWarning(UserWarning):
-
-    """Visible deprecation warning.
-    By default, python will not show deprecation warnings, so this class
-    provides a visible one.
-
-    """
-    pass
